refactor(mn): allow subjectNodeType to be a string
It’s not required required to have the actual node type instance,
and in some cases it’s not available. In those cases generateContent()
can be passed “inline” or “block” to indicate what kind of content
is considered valid.

(If subjectNodeType is a node type instance, generateContent()
will run PM’s content validation function that should do that
automatically.)